refactor loginPage with splitting components
